redesign of interviews and buttons
